Moved type contracts to the common package.
